+++ cpp/a3/src/TsPusher.h
@@ -92,7 +92,7 @@ public:
 	}
 
 	#pragma region 通过 Stream 继承
-	void Write(uint8_t *src_buf, int64_t offset, int64_t count) override
+	void Write(uint8_t *src_buf, int32_t offset, int32_t count) override
 	{
 		_mstream.Write(src_buf, offset, count);
 	}
@@ -102,7 +102,7 @@ public:
 	bool CanSeek() override;
 	int64_t Length() override;
 	void SetLength(int64_t value) override;
-	int64_t Read(uint8_t *dst_buf, int64_t offset, int64_t count) override;
+	int32_t Read(uint8_t *dst_buf, int32_t offset, int32_t count) override;
 	void Flush() override;
 	void Close() override;
 	int64_t Position() override;
