+++ cpp/jccpp/include/jccpp/stream/FileStream.h
@@ -46,8 +46,8 @@ public:
 	int64_t Length() override;
 	void SetLength(int64_t value) override;
 
-	int64_t Read(uint8_t *buffer, int64_t offset, int64_t count) override;
-	void Write(uint8_t *buffer, int64_t offset, int64_t count) override;
+	int32_t Read(uint8_t *buffer, int32_t offset, int32_t count) override;
+	void Write(uint8_t *buffer, int32_t offset, int32_t count) override;
 
 	int64_t Position() override;
 	void SetPosition(int64_t value) override;
