+++ cpp/jccpp/include/jccpp/stream/JoinedStream.cpp
@@ -52,7 +52,7 @@ void JoinedStream::SetLength(int64_t value)
 	throw jc::NotSupportedException();
 }
 
-int64_t JoinedStream::Read(uint8_t *dst_buf, int64_t offset, int64_t count)
+int32_t JoinedStream::Read(uint8_t *dst_buf, int32_t offset, int32_t count)
 {
 	while (1)
 	{
@@ -79,7 +79,7 @@ int64_t JoinedStream::Read(uint8_t *dst_buf, int64_t offset, int64_t count)
 	}
 }
 
-void JoinedStream::Write(uint8_t *src_buf, int64_t offset, int64_t count)
+void JoinedStream::Write(uint8_t *src_buf, int32_t offset, int32_t count)
 {
 	throw jc::NotSupportedException();
 }
